requirements.txt
numpy==1.24.2
opencv_python==4.7.0.72
PyAudio==0.2.13
PyAutoGUI==0.9.53
pydub==0.25.1
pyscreenshot==3.1
pyserial==3.5
pytesseract==0.3.10
SpeechRecognition==3.10.0
